requirements.txt
# Creating the virtual environment
# python -m venv venv

# Activting the venv 
# venv\Scripts\activate

# Installing the flask
# pip install flask

# Installing sql database library
# pip install flask_sqlalchemy

# connecting with the mysql databse
# pip install mysqlclient

# Installing QR
# pip install qrcode

# Installing jwt-extended
# pip install flask-jwt-extended

# Installing python .env
# pip install python-dotenv

# Installing flask cors for ngrok integration
# pip install flask-cors

#Installing cryptography
#pip install cryptography

# Intsalling flask restful
# pip install flask_restful

# Installing fuzzywuzz for user search suggestion and matching
# pip install fuzzywuzzy

# Installin python-Levenshtein to make search suggestion and matching faster
# pip install python-Levenshtein


#pip install requests

#pip install apscheduler

blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.1
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug~=3.0.1

requests~=2.31.0
cryptography~=41.0.7
fuzzywuzzy~=0.18.0
pytest~=7.4.4